Sentiment_Analysis/sa_databroker/requirements.txt
Bootstrap-Flask==1.5.2
Flask==1.1.2
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.14.3
google==3.0.0
google-auth==1.35.0
google-auth-oauthlib==0.4.6
googleapis-common-protos==1.53.0
grpcio==1.38.0
grpcio-tools==1.38.0
Jinja2==2.11.3
protobuf==3.16.0
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==1.3.0
threadpoolctl==2.2.0
WTForms==2.3.3
